Refactor room filtering and mapping in getRooms()
The getRooms() function has been refactored to filter and map the rooms array more efficiently. Now, it filters the rooms that start with 'room' and have less than 3 players. It then maps the filtered rooms to only include their names.

Also, update endOfTournament() to remove room and players

The endOfTournament() function has been updated to remove the room from the roomData object and remove all players from the room. This ensures that the room and its associated data are properly cleaned up after the tournament ends.

Update room-select options in socket.on('getRooms')

The options in the room-select dropdown menu have been updated based on the received data from the server. If there are new rooms available, they are added as options. If there are no rooms available, only the "Créer nouvelle room" option is displayed.